Refactor gRPC service definition and add proto compilation script
