+++ pdf-grpc-server.py
@@ -5,7 +5,6 @@ import tempfile
 from concurrent import futures
 from typing import Dict, Any
 import json
-import subprocess
 
 import grpc
 from google.protobuf import json_format, empty_pb2
@@ -23,65 +22,9 @@ logging.basicConfig(
 # Load environment variables
 load_dotenv()
 
-# Define the proto file for our gRPC service
-proto_content = """
-syntax = "proto3";
-
-package pdf_processor;
-
-import "google/protobuf/empty.proto";
-import "google/protobuf/struct.proto";
-
-service PdfProcessor {
-    rpc ProcessPdf(PdfRequest) returns (PdfStructResponse) {}
-    rpc HealthCheck(google.protobuf.Empty) returns (HealthCheckResponse) {}
-}
-
-message PdfRequest {
-    bytes pdf_data = 1;
-    string filename = 2;
-}
-
-message PdfStructResponse {
-    google.protobuf.Struct data = 1;
-    string message = 2;
-    bool success = 3;
-}
-
-message HealthCheckResponse {
-    string status = 1;
-}
-"""
-
-def compile_proto():
-    """Compile the proto file and ensure it succeeds"""
-    # Write the proto file to disk
-    with open('pdf_processor.proto', 'w') as f:
-        f.write(proto_content)
-    
-    # Compile using subprocess to check for errors
-    try:
-        result = subprocess.run(
-            'python -m grpc_tools.protoc -I. --python_out=. --grpc_python_out=. pdf_processor.proto',
-            shell=True,
-            check=True,
-            text=True,
-            capture_output=True
-        )
-        logging.info("Proto compilation succeeded")
-        return True
-    except subprocess.CalledProcessError as e:
-        logging.error(f"Proto compilation failed: {e.stderr}")
-        return False
-
-# Compile the proto file
-if not compile_proto():
-    logging.error("Failed to compile proto file. Exiting.")
-    exit(1)
-
 # Import the generated gRPC modules
 import pdf_processor_pb2
-import pdf_processor_pb2_grpc
+import pdf_processor_pb2_grpcW
 
 # Constants
 MAX_MESSAGE_LENGTH = 100 * 1024 * 1024  # 100MB
